functions/predict/requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.2.0
cachetools==4.2.2
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==8.0.1
cloudevents==1.2.0
decorator==5.0.9
deprecation==2.1.0
Flask==2.0.1
fsspec==2021.5.0
functions-framework==2.2.0
gcsfs==2021.5.0
google-api-core==1.28.0
google-auth==1.30.0
google-auth-oauthlib==0.4.4
google-cloud-core==1.6.0
google-cloud-storage==1.38.0
google-crc32c==1.1.2
google-resumable-media==1.3.0
googleapis-common-protos==1.53.0
gunicorn==20.1.0
idna==2.10
itsdangerous==2.0.1
Jinja2==3.0.1
joblib==1.0.1
MarkupSafe==2.0.1
multidict==5.1.0
numpy==1.20.3
oauthlib==3.1.0
packaging==20.9
pandas==1.2.4
protobuf==3.17.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
scikit-learn==0.23.2
scipy==1.6.3
six==1.16.0
threadpoolctl==2.1.0
typing-extensions==3.10.0.0
urllib3==1.26.4
watchdog==1.0.2
Werkzeug==2.0.1
yarl==1.6.3
